Move gzip/gunzip methods into Buffers class
